Removed unnecessary permissions from example app Improved in-code documentation Fixed spelling and grammar mistakes Refactored internal structure Removed unecessary code
